requirements.txt
# Solana Core
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0

# Web3 & Blockchain
base58==2.1.1
construct==2.10.68

# API & HTTP
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Social Media APIs
TikTokApi==6.3.1
tweepy==4.14.0

# AI & NLP (for sentiment analysis)
openai==1.6.1
anthropic==0.8.1

# Database
sqlalchemy==2.0.23
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Monitoring
sentry-sdk==1.39.1
prometheus-client==0.19.0
